remove GCCCCompiler, GCCCLinker, etc
