+++ src/python/pants/backend/python/subsystems/BUILD
@@ -6,6 +6,8 @@ python_library(
     '3rdparty/python:future',
     '3rdparty/python:pex',
     '3rdparty/python:setuptools',
+    'src/python/pants/engine:rules',
+    'src/python/pants/engine:selectors',
     'src/python/pants/option',
     'src/python/pants/subsystem',
   ],
