+++ src/python/pants/backend/python/subsystems/python_native_code.py
@@ -4,18 +4,27 @@
 
 from __future__ import absolute_import, division, print_function, unicode_literals
 
+import os
 from builtins import str
 from collections import defaultdict
 
+from wheel.install import WheelFile
+
+from pants.backend.native.config.environment import (CppToolchain, CToolchain, LLVMCppToolchain,
+                                                     LLVMCToolchain, Platform)
 from pants.backend.native.subsystems.native_toolchain import NativeToolchain
 from pants.backend.native.targets.native_library import NativeLibrary
 from pants.backend.python.subsystems.python_setup import PythonSetup
 from pants.backend.python.targets.python_binary import PythonBinary
 from pants.backend.python.targets.python_distribution import PythonDistribution
+from pants.backend.python.tasks.pex_build_util import resolve_multi
 from pants.base.exceptions import IncompatiblePlatformsError
+from pants.engine.rules import RootRule, rule
+from pants.engine.selectors import Get, Select
 from pants.subsystem.subsystem import Subsystem
 from pants.util.memo import memoized_property
-from pants.util.objects import Exactly
+from pants.util.objects import Exactly, datatype
+from pants.util.strutil import create_path_env_var, safe_shlex_join
 
 
 class PythonNativeCode(Subsystem):
@@ -129,3 +138,155 @@ class PythonNativeCode(Subsystem):
       'native code. Please ensure that the platform arguments in all relevant targets and build '
       'options are compatible with the current platform. Found targets for platforms: {}'
       .format(str(platforms_with_sources)))
+
+
+@rule(CToolchain, [Select(PythonNativeCode)])
+def select_c_toolchain_for_local_dist_compilation(python_native_code):
+  llvm_c_toolchain = yield Get(LLVMCToolchain, NativeToolchain, python_native_code.native_toolchain)
+  yield llvm_c_toolchain.c_toolchain
+
+
+@rule(CppToolchain, [Select(PythonNativeCode)])
+def select_cpp_toolchain_for_local_dist_compilation(python_native_code):
+  llvm_cpp_toolchain = yield Get(
+    LLVMCppToolchain, NativeToolchain, python_native_code.native_toolchain)
+  yield llvm_cpp_toolchain.cpp_toolchain
+
+
+class SetupPyNativeTools(datatype([
+    ('c_toolchain', CToolchain),
+    ('cpp_toolchain', CppToolchain),
+    ('platform', Platform),
+])):
+  """The native tools needed for a setup.py invocation.
+
+  This class exists because `SetupPyExecutionEnvironment` is created manually, one per target.
+  """
+
+
+# TODO: could this kind of @rule be automatically generated?
+@rule(SetupPyNativeTools, [Select(CToolchain), Select(CppToolchain), Select(Platform)])
+def get_setup_py_native_tools(c_toolchain, cpp_toolchain, platform):
+  yield SetupPyNativeTools(
+    c_toolchain=c_toolchain,
+    cpp_toolchain=cpp_toolchain,
+    platform=platform)
+
+
+class SetupRequiresSiteDir(datatype(['site_dir'])): pass
+
+
+# TODO: This could be formulated as an @rule if targets and `PythonInterpreter` are made available
+# to the v2 engine.
+def ensure_setup_requires_site_dir(reqs_to_resolve, interpreter, site_dir,
+                                   platforms=None):
+  if not reqs_to_resolve:
+    return None
+
+  setup_requires_dists = resolve_multi(interpreter, reqs_to_resolve, platforms, None)
+
+  # FIXME: there's no description of what this does or why it's necessary.
+  overrides = {
+    'purelib': site_dir,
+    'headers': os.path.join(site_dir, 'headers'),
+    'scripts': os.path.join(site_dir, 'bin'),
+    'platlib': site_dir,
+    'data': site_dir
+  }
+
+  # The `python_dist` target builds for the current platform only.
+  # FIXME: why does it build for the current platform only?
+  for obj in setup_requires_dists['current']:
+    wf = WheelFile(obj.location)
+    wf.install(overrides=overrides, force=True)
+
+  return SetupRequiresSiteDir(site_dir)
+
+
+# TODO: It might be pretty useful to have an Optional TypeConstraint.
+class SetupPyExecutionEnvironment(datatype([
+    # If None, don't set PYTHONPATH in the setup.py environment.
+    'setup_requires_site_dir',
+    # If None, don't execute in the toolchain environment.
+    'setup_py_native_tools',
+])):
+
+  _SHARED_CMDLINE_ARGS = {
+    'darwin': lambda: [
+      '-mmacosx-version-min=10.11',
+      '-Wl,-dylib',
+      '-undefined',
+      'dynamic_lookup',
+    ],
+    'linux': lambda: ['-shared'],
+  }
+
+  def as_environment(self):
+    ret = {}
+
+    if self.setup_requires_site_dir:
+      ret['PYTHONPATH'] = self.setup_requires_site_dir.site_dir
+
+    # FIXME(#5951): the below is a lot of error-prone repeated logic -- we need a way to compose
+    # executables more hygienically. We should probably be composing each datatype's members, and
+    # only creating an environment at the very end.
+    native_tools = self.setup_py_native_tools
+    if native_tools:
+      # TODO: an as_tuple() method for datatypes could make this destructuring cleaner!
+      plat = native_tools.platform
+      c_toolchain = native_tools.c_toolchain
+      c_compiler = c_toolchain.c_compiler
+      c_linker = c_toolchain.c_linker
+
+      cpp_toolchain = native_tools.cpp_toolchain
+      cpp_compiler = cpp_toolchain.cpp_compiler
+      cpp_linker = cpp_toolchain.cpp_linker
+
+      all_path_entries = (
+        c_compiler.path_entries +
+        c_linker.path_entries +
+        cpp_compiler.path_entries +
+        cpp_linker.path_entries)
+      ret['PATH'] = create_path_env_var(all_path_entries)
+
+      all_library_dirs = (
+        c_compiler.library_dirs +
+        c_linker.library_dirs +
+        cpp_compiler.library_dirs +
+        cpp_linker.library_dirs)
+      joined_library_dirs = create_path_env_var(all_library_dirs)
+      dynamic_lib_env_var = plat.resolve_platform_specific({
+        'darwin': lambda: 'DYLD_LIBRARY_PATH',
+        'linux': lambda: 'LD_LIBRARY_PATH',
+      })
+      ret[dynamic_lib_env_var] = joined_library_dirs
+
+      all_linking_library_dirs = (c_linker.linking_library_dirs + cpp_linker.linking_library_dirs)
+      ret['LIBRARY_PATH'] = create_path_env_var(all_linking_library_dirs)
+
+      all_include_dirs = c_compiler.include_dirs + cpp_compiler.include_dirs
+      ret['CPATH'] = create_path_env_var(all_include_dirs)
+
+      all_cflags_for_platform = plat.resolve_platform_specific({
+        'darwin': lambda: ['-mmacosx-version-min=10.11', '-nostdinc++'],
+        'linux': lambda: ['-nostdinc++'],
+      })
+      ret['CFLAGS'] = safe_shlex_join(all_cflags_for_platform)
+
+      ret['CC'] = c_compiler.exe_filename
+      ret['CXX'] = cpp_compiler.exe_filename
+      ret['LDSHARED'] = cpp_linker.exe_filename
+
+      all_new_ldflags = plat.resolve_platform_specific(self._SHARED_CMDLINE_ARGS)
+      ret['LDFLAGS'] = safe_shlex_join(all_new_ldflags)
+
+    return ret
+
+
+def create_python_native_code_rules():
+  return [
+    select_c_toolchain_for_local_dist_compilation,
+    select_cpp_toolchain_for_local_dist_compilation,
+    get_setup_py_native_tools,
+    RootRule(PythonNativeCode),
+  ]
